app/requirements.txt
bcrypt==4.0.1
bidict==0.22.1
blinker==1.6.2
click==8.1.3
et-xmlfile==1.1.0
Flask==2.3.2
Flask-Cors==4.0.0
Flask-SocketIO==5.3.4
Flask-SQLAlchemy==3.0.5
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
openpyxl==3.1.2
PyJWT==2.7.0
python-dotenv==1.0.0
python-engineio==4.4.1
python-socketio==5.8.0
schedule==1.2.0
SQLAlchemy==2.0.17
typing_extensions==4.7.0
Werkzeug==2.3.6
